refactor(tests): split OCCTDrawingTests.swift by @Suite (#1299)
OCCTDrawingTests.swift carried 2780 lines across 33 @Suite structs. Split into
one file per suite, named after the struct, matching the sibling Issue*Tests.swift
files already in Tests/OCCTDrawingTests/. The one file-scope helper (SIMD3.normalized,
used by no suite in this file, per-target convention per CLAUDE.md's Test Layout)
moves to a new DrawingTestFixtures.swift, matching the Tests/OCCTShapeHealingTests/
ShapeHealingTestFixtures.swift precedent. The one single-suite helper (RecordingSink)
moves with its suite (DrawingTransformUnificationTests.swift), private -> fileprivate.

Pure move: no assertion, tolerance, or test body changed. @Test count unchanged
(197 before and after). OCCTDrawingTests.swift is now empty and removed.

Also fixes a stale cross-reference in Issue1173ArrowheadTriangleGeometryTests.swift's
own comment, which pointed at OCCTDrawingTests.swift for cuttingPlaneRoundtrip
(now on DrawingAppendTests.swift).

Closes #1299
